optimize/eval: address Copilot review (include cycle, generic inner_pow, empty BatchPolicy)
Three robustness fixes from the PR #559 Copilot review:

- Break the single_term.hpp <-> cost_model.hpp include cycle. The shared DP
  helpers (cost counters, OptRes/EvalSequence, init_results) move into a new
  single_term_detail.hpp; cost_model.hpp includes that instead of single_term.hpp,
  and single_term.hpp includes cost_model.hpp before defining the public
  single_term_opt<Metric> entry points. cost_model.hpp is now standalone-
  includable (verified by compiling a TU that includes it first). Register
  cost_model.hpp (previously unlisted) and single_term_detail.hpp in CMakeLists.

- inner_aware_volume() no longer requires its InnerPow template argument to be
  bool-convertible: an option_engaged() helper reports the engaged state of
  std::function/function pointers and treats any other callable (e.g. a bare
  lambda) as always engaged, so callers may pass a lambda for inner_pow.

- make_evaluator() defensively declines batching when BatchPolicy::is_batchable_index
  or batch_target_size is empty (per the BatchPolicy docs), instead of forwarding
  an empty std::function that would throw std::bad_function_call at eval time.

Full unit suite unchanged: 6375 assertions, 56 cases.
