+++ SeQuant/core/optimize/single_term_detail.hpp
@@ -1,0 +1,480 @@
+#ifndef SEQUANT_CORE_OPTIMIZE_SINGLE_TERM_DETAIL_HPP
+#define SEQUANT_CORE_OPTIMIZE_SINGLE_TERM_DETAIL_HPP
+
+// Shared single-term-optimization detail: cost counters, OptRes/EvalSequence,
+// init_results, and the subset/bipartition DP scaffolding. Split out of
+// single_term.hpp so that cost_model.hpp can include just these helpers
+// without the public single_term_opt<Metric> definitions (which themselves
+// depend on cost_model.hpp). This breaks the single_term.hpp <-> cost_model.hpp
+// include cycle and lets cost_model.hpp be included standalone.
+
+#include <SeQuant/core/algorithm.hpp>
+#include <SeQuant/core/container.hpp>
+#include <SeQuant/core/expr.hpp>
+#include <SeQuant/core/optimize/options.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
+#include <SeQuant/core/tensor_network.hpp>
+#include <SeQuant/core/utility/indices.hpp>
+#include <SeQuant/core/utility/macros.hpp>
+#include <SeQuant/external/bliss/graph.hh>
+
+#include <range/v3/algorithm/find.hpp>
+#include <range/v3/view/concat.hpp>
+
+#include <cstdlib>
+#include <iostream>
+#include <range/v3/view/filter.hpp>
+#include <range/v3/view/indirect.hpp>
+#include <range/v3/view/move.hpp>
+#include <range/v3/view/reverse.hpp>
+#include <range/v3/view/transform.hpp>
+
+#include <algorithm>
+#include <bit>
+#include <functional>
+#include <limits>
+#include <type_traits>
+
+namespace sequant::opt {
+
+template <typename F>
+concept has_index_extent = std::is_invocable_r_v<size_t, F, Index const&>;
+
+namespace detail {
+
+/// \brief Whether an "option-like" callable is engaged.
+///
+/// \c std::function and function pointers report their non-empty state via
+/// contextual conversion to bool; any other callable (e.g. a plain lambda) has
+/// no empty state and is always engaged. This lets cost helpers accept either a
+/// (possibly empty) \c std::function or a bare lambda for optional callables
+/// such as \c inner_pow, instead of requiring a bool-convertible type.
+template <typename F>
+constexpr bool option_engaged(F const& f) {
+  if constexpr (std::is_constructible_v<bool, F const&>)
+    return static_cast<bool>(f);
+  else
+    return true;
+}
+
+/// \brief Composite-aware extent product of a \ref tot_indices split.
+///
+/// Multiplies the extents of the outer indices (via \p ixex) by the extents of
+/// the inner (CSV/PNO tensor-of-tensor) composites. Shared by all cost counters
+/// (\ref flops_counter, \ref memsize_counter, \ref footprint_counter) so every
+/// objective sizes CSV/PNO composites identically.
+///
+/// \param tot_idxs a \ref tot_indices split (\c outer plus \c inner
+/// composites). \param ixex     invocable mapping an Index to its extent (sizes
+/// \c outer and,
+///                 when \p inner_pow is empty, \c inner too).
+/// \param inner_pow optional invocable <tt>(composite, k) -> double</tt> giving
+///                 the size of a composite that shares its proto-index set with
+///                 \c k composites in the same tensor (the k-th power mean of
+///                 the per-pair domain). When non-empty, composites are grouped
+///                 by proto-index set and each member of a k-group is sized by
+///                 \c inner_pow(composite, k); the outer \c nocc^N times the
+///                 group product then equals the true block-sparse volume
+///                 \c Sum_pairs d^k.
+/// \return the extent product (element count).
+/// \note When \p inner_pow is empty, composites fall back to \p ixex (i.e.
+///       \c k=1), reproducing a single grid-averaged extent per composite,
+///       which under-counts multi-composite tensors: for heavy-tailed domains
+///       the mean of d^k far exceeds the k-th power of the mean of d.
+template <typename Tot, typename Ixex, typename InnerPow>
+double inner_aware_volume(Tot const& tot_idxs, Ixex const& ixex,
+                          InnerPow const& inner_pow) {
+  double mem = ranges::accumulate(tot_idxs.outer, 1., std::multiplies{}, ixex);
+  if (option_engaged(inner_pow)) {
+    for (auto const& c : tot_idxs.inner) {
+      std::size_t k = 0;
+      for (auto const& o : tot_idxs.inner)
+        if (o.proto_indices() == c.proto_indices()) ++k;
+      mem *= inner_pow(c, k);
+    }
+  } else {
+    mem = ranges::accumulate(tot_idxs.inner, mem, std::multiplies{}, ixex);
+  }
+  return mem;
+}
+
+/// \brief Cost function returning the flop count of a binary tensor
+/// contraction.
+///
+/// The returned callable computes the product of extents over the union of
+/// indices on \c lhs, \c rhs, and \c result. A product of 1 (scalar
+/// contraction) is reported as zero flops.
+///
+/// \param ixex Invocable mapping an Index to its extent.
+/// \param inner_pow Optional k-aware CSV/PNO composite extent; see
+///        \ref inner_aware_volume. Empty (default) sizes composites by \p ixex.
+/// \return A callable <tt>(lhs, rhs, result) -> double</tt> yielding the
+/// flop count of the contraction.
+template <typename InnerPow = std::function<double(Index const&, std::size_t)>>
+auto flops_counter(has_index_extent auto&& ixex, InnerPow inner_pow = {}) {
+  return [ixex = std::forward<decltype(ixex)>(ixex),
+          inner_pow = std::move(inner_pow)](
+             meta::range_of<Index> auto const& lhs,
+             meta::range_of<Index> auto const& rhs,
+             meta::range_of<Index> auto const& result) -> double {
+    using ranges::views::concat;
+    // <IndexSet> is required here: concatenating the three operands repeats
+    // every contracted/shared index, so it must be deduplicated before taking
+    // the extent product (cf. memsize_counter, which processes each operand
+    // separately and so can use the default vector container).
+    auto tot_idxs = tot_indices<IndexSet>(concat(lhs, rhs, result));
+    double total_flops = inner_aware_volume(tot_idxs, ixex, inner_pow);
+    // A product of exactly 1. means the index set was empty (the accumulation
+    // init value), i.e. a scalar contraction => zero flops. Extents are
+    // integer-valued, so this equality is exact.
+    return total_flops == 1. ? 0. : total_flops;
+  };
+}
+
+/// \brief Cost function returning the total memory footprint of a binary
+/// tensor contraction.
+///
+/// The returned callable sums, over \c lhs, \c rhs, and \c result, the
+/// product of extents of each operand's indices. Operands whose extent
+/// product is 1 contribute zero.
+///
+/// \param ixex Invocable mapping an Index to its extent.
+/// \param inner_pow Optional k-aware CSV/PNO composite extent; see
+///        \ref inner_aware_volume. Empty (default) sizes composites by \p ixex.
+/// \return A callable <tt>(lhs, rhs, result) -> double</tt> yielding the
+/// summed memory size of the three operands.
+template <typename InnerPow = std::function<double(Index const&, std::size_t)>>
+auto memsize_counter(has_index_extent auto&& ixex, InnerPow inner_pow = {}) {
+  return [ixex = std::forward<decltype(ixex)>(ixex),
+          inner_pow = std::move(inner_pow)](
+             meta::range_of<Index> auto const& lhs,
+             meta::range_of<Index> auto const& rhs,
+             meta::range_of<Index> auto const& result) -> double {
+    double total_mem{0};
+    // Each operand is sized independently, so the default (vector) container of
+    // tot_indices suffices -- a single operand's index list has no duplicates,
+    // unlike the concatenated set flops_counter must dedup.
+    for (auto&& tot_idxs :
+         {tot_indices(lhs), tot_indices(rhs), tot_indices(result)}) {
+      double mem = inner_aware_volume(tot_idxs, ixex, inner_pow);
+      // mem == 1. means this operand had no indices (the accumulation init
+      // value), i.e. a scalar; it contributes no memory. Same exact-equality
+      // convention as flops_counter above.
+      if (mem != 1.) total_mem += mem;
+    }
+    return total_mem;
+  };
+}
+
+/// \brief Cost function returning the storage footprint (element count) of a
+/// single result tensor.
+///
+/// Sizes the result's outer indices by \p ixex and its inner (CSV/PNO
+/// tensor-of-tensor) composites by \p inner_pow (see \ref inner_aware_volume).
+/// Used both as the peak objectives' per-subset footprint (\ref
+/// subset_footprints) and as the additive objectives' per-intermediate
+/// footprint penalty (see OptimizeOptions::footprint_weight). A scalar (no
+/// indices) contributes zero.
+///
+/// \param ixex Invocable mapping an Index to its extent.
+/// \param inner_pow Optional k-aware CSV/PNO composite extent; see
+///        \ref inner_aware_volume. Empty (default) sizes composites by \p ixex,
+///        which under-counts multi-composite tensors.
+/// \return A callable <tt>(result) -> double</tt> yielding the element count of
+/// the result tensor.
+template <typename InnerPow = std::function<double(Index const&, std::size_t)>>
+auto footprint_counter(has_index_extent auto&& ixex, InnerPow inner_pow = {}) {
+  return [ixex = std::forward<decltype(ixex)>(ixex),
+          inner_pow = std::move(inner_pow)](
+             meta::range_of<Index> auto const& result) -> double {
+    double mem = inner_aware_volume(tot_indices(result), ixex, inner_pow);
+    return mem == 1. ? 0. : mem;
+  };
+}
+
+///
+/// Represents a result of optimization on a range of expressions
+/// for a binary evaluation
+///
+struct OptRes {
+  /// Free indices remaining upon evaluation
+  IndexSet indices;
+
+  /// The operations count of evaluation
+  double ops;
+
+  /// The evaluation sequence
+  EvalSequence sequence;
+
+  /// Bitmask splits that resulted into this OptRes
+  size_t lp = 0;
+  size_t rp = 0;
+
+  /// unique canonical subnets in the optimal tree for this bitmask
+  container::vector<size_t> subnets;
+};
+
+/// \brief Footprint (dense element count) of every subset's result tensor.
+///
+/// \c S[n] is the product of extents of the open indices of subset \c n
+/// (those remaining after contracting the tensors in \c n, given \c tidxs as
+/// the final target indices). \c S[0] (empty subset) and any scalar result are
+/// 0. Shared by the peak DP and its tests so both agree on per-subset sizes.
+///
+/// \param network The TensorNetwork.
+/// \param tidxs   Target (open) indices of the network.
+/// \param idxsz   Callable mapping an Index to its extent.
+/// \param inner_pow Optional k-aware CSV/PNO composite extent forwarded to
+///        \ref footprint_counter; see \ref inner_aware_volume.
+/// \return \c S[n] = footprint of subset \c n's result tensor.
+template <typename TIdxs, typename IdxToSz>
+container::vector<double> subset_footprints(
+    TensorNetwork const& network, TIdxs const& tidxs, IdxToSz&& idxsz,
+    std::function<double(Index const&, std::size_t)> const& inner_pow = {}) {
+  container::vector<OptRes> results((size_t{1} << network.tensors().size()));
+  init_results(network, tidxs, results);
+  auto fp = footprint_counter(std::forward<IdxToSz>(idxsz), inner_pow);
+  container::vector<double> S(results.size(), 0.0);
+  for (size_t n = 0; n < results.size(); ++n)
+    S[n] = (n == 0) ? 0.0 : fp(results[n].indices);
+  return S;
+}
+
+/// \brief Collects the distinct batchable indices (in appearance order) across
+/// all tensors in \p network.
+///
+/// Iterates over every tensor slot (bra, ket, and aux) and appends an index to
+/// the result the first time it is seen and \p is_batchable returns true for
+/// it.  The returned list assigns each index a stable bit position: index at
+/// position \c k is bit \c k of a sliced-set bitmask \c B.
+///
+/// \param network  The TensorNetwork to scan.
+/// \param is_batchable  Predicate returning true for indices in a batchable
+///        space (e.g. a DF/RI auxiliary space).
+/// \return Ordered, deduplicated list of batchable indices.
+inline container::vector<Index> batchable_index_list(
+    TensorNetwork const& network,
+    std::function<bool(Index const&)> const& is_batchable) {
+  container::vector<Index> aux;
+  if (!is_batchable) return aux;
+  for (auto&& t : network.tensors()) {
+    auto tp = std::dynamic_pointer_cast<Tensor>(t);
+    for (auto&& ix : ranges::views::concat(tp->bra(), tp->ket(), tp->aux()))
+      if (is_batchable(ix) && ranges::find(aux, ix) == ranges::end(aux))
+        aux.push_back(ix);
+  }
+  return aux;
+}
+
+/// \brief Footprint tables for every sliced-set B of batchable indices.
+///
+/// Returns a vector of \c 2^m tables, where \c m = \c aux_list.size(). Table
+/// \c B is the result of \ref subset_footprints evaluated with an extent
+/// function that replaces the full extent of \c aux_list[k] with
+/// \c min(full_extent, batch) whenever bit \c k is set in \c B.
+///
+/// \param network    The TensorNetwork.
+/// \param tidxs      Target (open) indices of the network.
+/// \param idxsz      Callable mapping an Index to its full extent.
+/// \param is_batchable  Predicate identifying batchable indices.
+/// \param batch_target_size  Per-index slice-size function: a sliced batchable
+///        index \c ix contributes
+///        min(full_extent, batch_target_size(aux_list[k])).
+/// \param aux_list   Ordered list of distinct batchable indices (as returned
+///        by \ref batchable_index_list).
+/// \param inner_pow Optional k-aware CSV/PNO composite extent forwarded to each
+///        per-\c B \ref subset_footprints call; see \ref inner_aware_volume.
+///        Orthogonal to slicing (composites are not the batchable aux indices).
+/// \return \c tables[B][n] = footprint of subset \c n under sliced-set \c B.
+template <typename TIdxs, typename IdxToSz>
+container::vector<container::vector<double>> sliced_footprints(
+    TensorNetwork const& network, TIdxs const& tidxs, IdxToSz&& idxsz,
+    std::function<bool(Index const&)> const& is_batchable,
+    std::function<std::size_t(Index const&)> const& batch_target_size,
+    container::vector<Index> const& aux_list,
+    std::function<double(Index const&, std::size_t)> const& inner_pow = {}) {
+  std::size_t const m = aux_list.size();
+  container::vector<container::vector<double>> tables(std::size_t{1} << m);
+  for (std::size_t B = 0; B < tables.size(); ++B) {
+    auto extent = [&, B](Index const& ix) -> std::size_t {
+      std::size_t e = idxsz(ix);
+      if (is_batchable && is_batchable(ix)) {
+        auto it = ranges::find(aux_list, ix);
+        if (it != ranges::end(aux_list)) {
+          std::size_t k =
+              static_cast<std::size_t>(it - ranges::begin(aux_list));
+          if (B & (std::size_t{1} << k))
+            return std::min(e, batch_target_size(ix));
+        }
+      }
+      return e;
+    };
+    tables[B] = subset_footprints(network, tidxs, extent, inner_pow);
+  }
+  return tables;
+}
+
+/// \brief Bitmask of volatile leaf tensors in \p network.
+///
+/// Bit \c i is set if tensor \c i (in \c network.tensors() order) satisfies
+/// \p is_volatile_leaf.  Returns 0 when the predicate is empty (no tensor is
+/// volatile, weighting disabled).
+///
+/// \param network           The TensorNetwork.
+/// \param is_volatile_leaf  Predicate identifying volatile leaf tensors; may
+///        be empty.
+/// \return Bitmask with bit i set iff tensor i is a volatile leaf.
+inline std::size_t leaf_volatile_mask(
+    TensorNetwork const& network,
+    std::function<bool(Tensor const&)> const& is_volatile_leaf) {
+  std::size_t mask = 0;
+  if (!is_volatile_leaf) return mask;
+  std::size_t i = 0;
+  for (auto&& t : network.tensors()) {
+    auto tp = std::dynamic_pointer_cast<Tensor>(t);
+    if (tp && is_volatile_leaf(*tp)) mask |= (std::size_t{1} << i);
+    ++i;
+  }
+  return mask;
+}
+
+struct SubNetHash {
+  size_t operator()(
+      TensorNetwork::SlotCanonicalizationMetadata const& data) const noexcept {
+    return data.hash_value();
+  }
+};
+
+struct SubNetEqual {
+  bool operator()(
+      TensorNetwork::SlotCanonicalizationMetadata const& left,
+      TensorNetwork::SlotCanonicalizationMetadata const& right) const {
+    return bliss::ConstGraphCmp::cmp(*left.graph, *right.graph) == 0;
+  }
+};
+
+/// \brief Seeds the DP table with per-subset open indices and singleton/empty
+/// ops counts.
+///
+/// For each subset \c i of the input tensors, computes the indices that remain
+/// open after evaluating that subset and stores them in \c results[i].indices.
+/// Initializes \c results[i].ops to 0 for empty and singleton subsets, and
+/// to \c max as a sentinel for subsets that will later be filled in by the DP.
+/// Singleton subsets also get their one-element \c sequence pre-populated.
+template <typename TIdxs>
+void init_results(TensorNetwork const& network, TIdxs const& tidxs,
+                  container::vector<OptRes>& results) {
+  using IndexContainer = decltype(OptRes::indices);
+  auto tensor_indices = network.tensors()          //
+                        | ranges::views::indirect  //
+                        | ranges::views::transform(slots);
+  auto imed_indices = subset_target_indices(tensor_indices, tidxs);
+  SEQUANT_ASSERT(ranges::distance(imed_indices) == ranges::distance(results));
+  for (size_t i = 0; i < results.size(); ++i) {
+    results[i].indices =
+        imed_indices[i] | ranges::views::move | ranges::to<IndexContainer>;
+    results[i].ops = std::popcount(i) > 1
+                         ? std::numeric_limits<decltype(OptRes::ops)>::max()
+                         : 0;
+    if (std::popcount(i) == 1)
+      results[i].sequence.emplace_back(std::countr_zero(i));
+    // else results[i].sequence is left uninitialized
+  }
+}
+
+struct SubnetMetadata {
+  /// meta_ids[n] is the canonical-subnet id of subset n, or
+  /// numeric_limits<size_t>::max() for subsets with popcount < 2.
+  container::vector<size_t> meta_ids;
+  /// Cost of evaluating one representative of each canonical subnet id,
+  /// indexed by id. Populated lazily during the DP.
+  container::vector<double> unique_meta_costs;
+};
+
+/// \brief Precomputes canonical-subnet identifiers for every subset of size
+/// >= 2 so that structurally equivalent subnetworks share a CSE id.
+///
+/// Builds a `TensorNetwork` for each subset, canonicalizes it, and assigns a
+/// dense integer id to each distinct canonical form. The returned
+/// `unique_meta_costs` is sized to the number of distinct ids and zero-filled;
+/// it is populated during the DP as each canonical subnet's optimal cost
+/// becomes known.
+///
+/// Side effect: `results[n].indices` may be reordered by `canonicalize_slots`.
+inline SubnetMetadata build_subnet_metadata(
+    TensorNetwork const& network, container::vector<OptRes>& results) {
+  SubnetMetadata out;
+  // Use max as sentinel for entries with popcount < 2 (singletons/empty),
+  // which are skipped below and never assigned a real meta ID.
+  out.meta_ids.resize(results.size(), std::numeric_limits<size_t>::max());
+  container::unordered_map<TensorNetwork::SlotCanonicalizationMetadata, size_t,
+                           SubNetHash, SubNetEqual>
+      meta_to_id;
+
+  for (size_t n = 0; n < results.size(); ++n) {
+    if (std::popcount(n) < 2) continue;
+    auto ts = bits::on_bits_index(n) | bits::sieve(network.tensors());
+    container::vector<ExprPtr> ts_expr;
+    for (auto&& t : ts)
+      ts_expr.emplace_back(std::dynamic_pointer_cast<Tensor>(t)->clone());
+
+    auto tn = TensorNetwork{ts_expr};
+    auto meta = tn.canonicalize_slots(
+        TensorCanonicalizer::cardinal_tensor_labels(), &results[n].indices);
+
+    auto [it, inserted] = meta_to_id.try_emplace(std::move(meta), 0);
+    if (inserted) it->second = meta_to_id.size() - 1;
+
+    out.meta_ids[n] = it->second;
+  }
+  out.unique_meta_costs.resize(meta_to_id.size(), 0.0);
+  return out;
+}
+
+/// \brief Per-subset bitmask of batchable indices that are OPEN in that subset.
+///
+/// For each subset \c n of the input tensors, bit \c k of \c open_aux[n] is set
+/// iff \c aux_list[k] is among the open (external) indices of subset \c n
+/// (those that remain after contracting \c n's tensors, with \c tidxs as the
+/// final targets). Used by the multi-mode batched DP and oracle to restrict the
+/// sliced-set context to indices actually open in a sized subset, so that table
+/// lookups (\ref sliced_footprints) are indexed consistently on both sides.
+///
+/// \param network   The TensorNetwork.
+/// \param tidxs     Target (open) indices of the network.
+/// \param aux_list  Ordered list of distinct batchable indices (as returned by
+///        \ref batchable_index_list); index \c k maps to bit \c k.
+/// \return \c open_aux[n] for every subset \c n.
+template <typename TIdxs>
+container::vector<std::size_t> subset_open_aux(
+    TensorNetwork const& network, TIdxs const& tidxs,
+    container::vector<Index> const& aux_list) {
+  container::vector<OptRes> results(
+      (std::size_t{1} << network.tensors().size()));
+  init_results(network, tidxs, results);
+  container::vector<std::size_t> open_aux(results.size(), 0);
+  for (std::size_t n = 0; n < results.size(); ++n) {
+    for (std::size_t k = 0; k < aux_list.size(); ++k) {
+      if (ranges::find(results[n].indices, aux_list[k]) !=
+          ranges::end(results[n].indices))
+        open_aux[n] |= (std::size_t{1} << k);
+    }
+  }
+  return open_aux;
+}
+
+/// Per-(subset, sliced-set) state for the multi-mode batched peak DP.
+/// Indexed in the flat table by \c n*(2^m)+B (subset \c n, sliced-set \c B).
+struct BatchedRes {
+  double peak = std::numeric_limits<double>::max();  // min peak for (n, B)
+  std::size_t lp = 0, rp = 0;  // winning bipartition (0 for singletons)
+  bool lp_first = true;        // winning evaluation order
+  std::size_t aprime = 0;  // winning subset of batchable indices sliced here
+  double flops = std::numeric_limits<double>::max();  // tie-break (min work
+                                                      // among equal-peak)
+};
+
+}  // namespace detail
+}  // namespace sequant::opt
+
+#endif  // SEQUANT_CORE_OPTIMIZE_SINGLE_TERM_DETAIL_HPP
